+++ src/anomalib/pipelines/benchmark/job.py
@@ -17,7 +17,8 @@ from rich.table import Table
 from anomalib.data import AnomalibDataModule
 from anomalib.engine import Engine
 from anomalib.models import AnomalyModule
-from anomalib.pipelines.components import Job, hide_output
+from anomalib.pipelines.components import Job
+from anomalib.utils.logging import hide_output
 
 logger = logging.getLogger(__name__)
