+++ src/anomalib/utils/config.py
@@ -3,9 +3,8 @@
 # Copyright (C) 2022-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-
 import logging
-from collections.abc import Sequence
+from collections.abc import Iterable, Sequence, ValuesView
 from pathlib import Path
 from typing import Any, cast
 
@@ -27,6 +26,34 @@ def _convert_nested_path_to_str(config: Any) -> Any:  # noqa: ANN401
     elif isinstance(config, Path | JSONArgparsePath):
         config = str(config)
     return config
+
+
+def to_nested_dict(config: dict) -> dict:
+    """Convert the flattened dictionary to nested dictionary.
+
+    Examples:
+        >>> config = {
+                "dataset.category": "bottle",
+                "dataset.image_size": 224,
+                "model_name": "padim",
+            }
+        >>> to_nested_dict(config)
+        {
+            "dataset": {
+                "category": "bottle",
+                "image_size": 224,
+            },
+            "model_name": "padim",
+        }
+    """
+    out: dict[str, Any] = {}
+    for key, value in config.items():
+        keys = key.split(".")
+        _dict = out
+        for k in keys[:-1]:
+            _dict = _dict.setdefault(k, {})
+        _dict[keys[-1]] = value
+    return out
 
 
 def to_yaml(config: Namespace | ListConfig | DictConfig) -> str:
@@ -76,6 +103,112 @@ def to_tuple(input_size: int | ListConfig) -> tuple[int, int]:
         msg = f"Expected either int or ListConfig, got {type(input_size)}"
         raise TypeError(msg)
     return ret_val
+
+
+def convert_valuesview_to_tuple(values: ValuesView) -> list[tuple]:
+    """Convert a ValuesView object to a list of tuples.
+
+    This is useful to get list of possible values for each parameter in the config and a tuple for values that are
+    are to be patched. Ideally this is useful when used with product.
+
+    Example:
+        >>> params = DictConfig({
+                "dataset.category": [
+                    "bottle",
+                    "cable",
+                ],
+                "dataset.image_size": 224,
+                "model_name": ["padim"],
+            })
+        >>> convert_to_tuple(params.values())
+        [('bottle', 'cable'), (224,), ('padim',)]
+        >>> list(itertools.product(*convert_to_tuple(params.values())))
+        [('bottle', 224, 'padim'), ('cable', 224, 'padim')]
+
+    Args:
+        values: ValuesView: ValuesView object to be converted to a list of tuples.
+
+    Returns:
+        list[Tuple]: List of tuples.
+    """
+    return_list = []
+    for value in values:
+        if isinstance(value, Iterable) and not isinstance(value, str):
+            return_list.append(tuple(value))
+        else:
+            return_list.append((value,))
+    return return_list
+
+
+def flatten_dict(config: dict, prefix: str = "") -> dict:
+    """Flatten the dictionary.
+
+    Examples:
+        >>> config = {
+                "dataset": {
+                    "category": "bottle",
+                    "image_size": 224,
+                },
+                "model_name": "padim",
+            }
+        >>> flatten_dict(config)
+        {
+            "dataset.category": "bottle",
+            "dataset.image_size": 224,
+            "model_name": "padim",
+        }
+    """
+    out = {}
+    for key, value in config.items():
+        if isinstance(value, dict):
+            out.update(flatten_dict(value, f"{prefix}{key}."))
+        else:
+            out[f"{prefix}{key}"] = value
+    return out
+
+
+def namespace_from_dict(container: dict) -> Namespace:
+    """Convert dictionary to Namespace recursively.
+
+    Examples:
+        >>> container = {
+                "dataset": {
+                    "category": "bottle",
+                    "image_size": 224,
+                },
+                "model_name": "padim",
+            }
+        >>> namespace_from_dict(container)
+        Namespace(dataset=Namespace(category='bottle', image_size=224), model_name='padim')
+    """
+    output = Namespace()
+    for k, v in container.items():
+        if isinstance(v, dict):
+            setattr(output, k, namespace_from_dict(v))
+        else:
+            setattr(output, k, v)
+    return output
+
+
+def dict_from_namespace(container: Namespace) -> dict:
+    """Convert Namespace to dictionary recursively.
+
+    Examples:
+        >>> from jsonargparse import Namespace
+        >>> ns = Namespace()
+        >>> ns.a = 1
+        >>> ns.b = Namespace()
+        >>> ns.b.c = 2
+        >>> dict_from_namespace(ns)
+        {'a': 1, 'b': {'c': 2}}
+    """
+    output = {}
+    for k, v in container.__dict__.items():
+        if isinstance(v, Namespace):
+            output[k] = dict_from_namespace(v)
+        else:
+            output[k] = v
+    return output
 
 
 def update_config(config: DictConfig | ListConfig | Namespace) -> DictConfig | ListConfig | Namespace:
